Effects, bugfix
1. Effects now both in misc class and .json
2. Inventory slot generation is now divided - misc generation in misc class
3. Bugfix

+++ animal.h
@@ -8,7 +8,7 @@ class Animal
 {
 public:
     Animal();
-    Animal(QString, QString, size_t, QMap<QString, size_t>);
+    Animal(QString, QString, size_t, QMap<QString, int>);
     QString name;
     QString type;
     size_t level;
